+++ src/base/regex.cpp
@@ -5,7 +5,8 @@
  * @link   http://chensoft.com
  */
 #include <chen/base/regex.hpp>
-#include <regex>
+#include <chen/base/str.hpp>
+#include <chen/base/num.hpp>
 
 using namespace chen;
 
@@ -13,24 +14,57 @@ using namespace chen;
 // regex
 bool regex::match(const std::string &text, const std::string &pattern)
 {
-    return std::regex_search(text, std::regex(pattern));
+    return regex::match(text, std::regex(pattern));
 }
 
-std::string regex::replace(const std::string &text,
-                           const std::string &pattern,
-                           const std::string &replace)
+bool regex::match(const std::string &text, const std::regex &regex)
 {
-    return std::regex_replace(text, std::regex(pattern), replace);
+    return std::regex_search(text, regex);
 }
 
-std::vector<std::string> regex::group(const std::string &text,
-                                      const std::string &pattern)
+std::string regex::replace(const std::string &text, const std::string &pattern, const std::string &replace)
+{
+    return regex::replace(text, std::regex(pattern), replace);
+}
+
+std::string regex::replace(const std::string &text, const std::regex &regex, const std::string &replace)
+{
+    return std::regex_replace(text, regex, replace);
+}
+
+std::string regex::expand(const std::string &text, const std::string &pattern, const std::string &expand)
+{
+    return regex::expand(text, std::regex(pattern), expand);
+}
+
+std::string regex::expand(const std::string &text, const std::regex &regex, const std::string &expand)
+{
+    std::smatch mt;
+
+    auto ret = expand;
+    auto idx = 0;
+
+    if (std::regex_search(text, mt, regex))
+    {
+        for (auto &str : mt)
+            chen::str::replace(ret, "$" + chen::num::str(idx++), str);
+    }
+
+    return ret;
+}
+
+std::vector<std::string> regex::group(const std::string &text, const std::string &pattern)
+{
+    return regex::group(text, std::regex(pattern));
+}
+
+std::vector<std::string> regex::group(const std::string &text, const std::regex &regex)
 {
     std::vector<std::string> ret;
 
     std::smatch mt;
 
-    if (std::regex_search(text, mt, std::regex(pattern)))
+    if (std::regex_search(text, mt, regex))
     {
         for (auto &str : mt)
             ret.emplace_back(str.str());